move rcraprofile to core app
