+++ server/apps/sites/models/__init__.py
@@ -1,3 +1,4 @@
+from ...core.models import RcraProfile
 from .contact_models import Address, Contact, RcraPhone, RcraStates
-from .profile_models import RcraProfile, RcraSitePermission
-from .site_models import RcraSite, RcraSiteType, Site, Role
+from .profile_models import RcraSitePermission
+from .site_models import RcraSite, RcraSiteType, Role, Site
